feat: whitelist becomes custom filters
Whitelist can now be set to Allow or Block and accepts regex as long as the first name starts and ends with '/'

Last Name and Server Name fields should be removed at a future date as well as a rewrite of most of the whitelist explanation. For now, let's give the translators a break.