fix: Disable logging outside of Debug mode ; fix custom emotes
Hunts and instances with lots of chat were spamming Log on every single message with far too much information. Logs are now disabled outside of Debug/Dry Run mode to avoid this performance hit.

In addition - Custom Emotes now have an override to always be shown if the player has enabled to show their own emotes in the emotes filters.