Use Dalamud window system
Using the Dalamud window system adds some benefits such as closing automatically when pressing Esc.

This also fixes an ImGui assert on ImGui.SetNextWindowSize, since you cannot combine ImGuiConds with |, it's meant to be used an enum instead of bitflags. I'm not entirely sure if Appearing or FirstUseEver is the desired value here, but I picked FirstUseEver since it's the closest to the current behaviour.