Fix DTR Entry (#106)
The "display current instance in DTR bar" setting currently doesn't seem to be working, and seems to occasionally be printing null pointer exceptions in the log.

It looks like this might've been broken in 55a3499 as part of the API 11 bump, where the DtrEntry variable no longer got set.

While I was at it, I also added using the `Hidden` attribute when the DTR entry is not shown, which bypasses some of the rendering, and also added a tooltip so when you hover over the instance it says "TidyChat", since this fix might mean that people who had this setting enabled might suddenly get a new instance number in the DTR bar and not realise where it is from.